Refactor; UX enhancements
