+++ changelog.md
@@ -4,63 +4,72 @@ All notable changes to this project will be documented in this file, starting wi
 
 ## \[1.0.7] - unreleased
 
-*   Added: Back to Login button on Register page
+- Fixed: Mobile Safari's floating bottom URL bar blocks sticky Search buttons.
+- Fixed: Credits without `position` terms infinitely loading
+- Added: Back to Login button on Register page
+- Added: Jump to Credits button (Edit Profile)
+- Improved: Cache policy on related skills
+- Improved: Changed website URL display to "View Website" to handle for long, ugly links.
+- Improved: Minor styling tweaks
+- Improved: Credit Save/Cancel buttons no longer rerendering unncessarily
+- Improved: Credits saved without `position` terms show a reminder on Edit screen
+- Refactored: EditCreditModal component, and merged contents of EditCreditView into it.
 
 ## \[1.0.6] - 2023-06-12 - 2023-06-13
 
-*   Improved: Mobile search UI - scroll to Jobs after selecting Department to indicate user should continue to scroll
-*   Improved: Unified style of React event, element, and type imports
-*   Added: Profile slug editing
-*   Fixed: Clear resume button restored
-*   Hotfixed: Social inputs not saving
-*   Hotfixed: Authentication refactor
-*   Bugs squashed, layouts fixed, quirks handled
+- Improved: Mobile search UI - scroll to Jobs after selecting Department to indicate user should continue to scroll
+- Improved: Unified style of React event, element, and type imports
+- Added: Profile slug editing
+- Fixed: Clear resume button restored
+- Hotfixed: Social inputs not saving
+- Hotfixed: Authentication refactor
+- Bugs squashed, layouts fixed, quirks handled
 
 ## \[1.0.5] - 2023-06-06
 
-*   Official Launcherino!
-*   Improved: Accessibility - Refactored CheckboxButton and RadioButton and respective groups to be simpler and more accessible
-*   Improved: Styling
-*   Improved: Mobile layouts and responsiveness
+- Official Launcherino!
+- Improved: Accessibility - Refactored CheckboxButton and RadioButton and respective groups to be simpler and more accessible
+- Improved: Styling
+- Improved: Mobile layouts and responsiveness
 
 ## \[1.0.4-beta] - 2023-06-05
 
-*   Added: Externally linked user taxonomy support (Partner Directories)
-*   Added: Email support button on Help page
-*   Improved: File/image uploading and removal UX
-*   Improved: Profile slugs! no more ID routing
-*   Removed: old AuthContext file (finally)
-*   Fixed: Socials not saving on profile edit
+- Added: Externally linked user taxonomy support (Partner Directories)
+- Added: Email support button on Help page
+- Improved: File/image uploading and removal UX
+- Improved: Profile slugs! no more ID routing
+- Removed: old AuthContext file (finally)
+- Fixed: Socials not saving on profile edit
 
 ## \[1.0.3-beta.2] - 2023-05-31
 
-*   Improved: Semantic versioning
-*   Improved: Header is now sticky
-*   Fixed: Search results no longer sometimes appear blank after navigating
+- Improved: Semantic versioning
+- Improved: Header is now sticky
+- Fixed: Search results no longer sometimes appear blank after navigating
 
 ## \[1.0.3beta] - 2023-05-30
 
-*   Added: "account already exists" message for account creation
-*   Improved: header mobile layout
-*   Improved: login mobile responsiveness
-*   Improved: checkbox/radio button fonts
-*   Fixed: Deleting last credit now correctly shows the credit's deletion
-*   Improved: Mobile profile edit layout
-*   Improved: Input colors
-*   Fixed: social network links, sanitization, and instructions
+- Added: "account already exists" message for account creation
+- Improved: header mobile layout
+- Improved: login mobile responsiveness
+- Improved: checkbox/radio button fonts
+- Fixed: Deleting last credit now correctly shows the credit's deletion
+- Improved: Mobile profile edit layout
+- Improved: Input colors
+- Fixed: social network links, sanitization, and instructions
 
 ## \[1.0.2beta] - 2023-05-24
 
-*   Added: Search by Name
-*   Added: ContentView component for displaying WP content
-*   Fixed: Search results now return a much longer list of results (pre-pagination support)
-*   Changed: Support email added to footer
+- Added: Search by Name
+- Added: ContentView component for displaying WP content
+- Fixed: Search results now return a much longer list of results (pre-pagination support)
+- Changed: Support email added to footer
 
 ## \[1.0.1beta] - 2023-05-16
 
-*   Added: Google reCAPTCHA v3 integration
+- Added: Google reCAPTCHA v3 integration
 
 ## \[1.0.0beta] - 2023-05-15
 
-*   Beta launch!
-*   Added: changelog
+- Beta launch!
+- Added: changelog
