+++ src/views/EditProfileView.tsx
@@ -618,19 +618,17 @@ export default function EditProfileView({ profile, profileLoading }: Props): JSX
 	const handleScrollToCredits = () => {
 		const credits = document.getElementById('credits');
 		if (credits) {
-			credits.scrollIntoView({ behavior: 'smooth' });
+			credits.scrollIntoView({ behavior: 'smooth', block: 'center', inline: 'nearest' });
 		}
 	};
 
 	return editProfile ? (
 		<form id='edit-profile' onSubmit={handleSubmit}>
-			<Heading variant='contentSubtitle' mt={2}>
-				Don't forget to{' '}
-				<Link onClick={handleScrollToCredits} title='Scroll to Credits'>
-					add your credits
-				</Link>{' '}
-				to appear in search results!
-			</Heading>
+			<Text variant='contentSubtitle' mt={2}>
+				<Button onClick={handleScrollToCredits} title='Scroll to Credits' textDecoration='none'>
+					Jump to Credits
+				</Button>
+			</Text>
 			<Flex
 				alignItems='center'
 				justifyContent='flex-end'
